requirement.txt
pandas
requests
ta
pandas_ta
google-auth
google-api-python-client
google-auth-oauthlib